+++ crates/node/builder/src/components/payload.rs
@@ -4,9 +4,11 @@ use crate::{BuilderContext, FullNodeTypes};
 use reth_basic_payload_builder::{BasicPayloadJobGenerator, BasicPayloadJobGeneratorConfig};
 use reth_chain_state::CanonStateSubscriptions;
 use reth_node_api::{NodeTypes, PayloadBuilderFor};
-use reth_payload_builder::{PayloadBuilderHandle, PayloadBuilderService};
+use reth_payload_builder::{PayloadBuilderHandle, PayloadBuilderService, PayloadServiceCommand};
 use reth_transaction_pool::TransactionPool;
 use std::future::Future;
+use tokio::sync::{broadcast, mpsc};
+use tracing::warn;
 
 /// A type that knows how to spawn the payload service.
 pub trait PayloadServiceBuilder<Node: FullNodeTypes, Pool: TransactionPool, EvmConfig>:
@@ -108,5 +110,46 @@ where
         ctx.task_executor().spawn_critical("payload builder service", Box::pin(payload_service));
 
         Ok(payload_service_handle)
+    }
+}
+
+/// A `NoopPayloadServiceBuilder` useful for node implementations that are not implementing
+/// validating/sequencing logic.
+#[derive(Debug, Clone, Copy, Default)]
+#[non_exhaustive]
+pub struct NoopPayloadServiceBuilder;
+
+impl<Node, Pool, Evm> PayloadServiceBuilder<Node, Pool, Evm> for NoopPayloadServiceBuilder
+where
+    Node: FullNodeTypes,
+    Pool: TransactionPool,
+    Evm: Send,
+{
+    async fn spawn_payload_builder_service(
+        self,
+        ctx: &BuilderContext<Node>,
+        _pool: Pool,
+        _evm_config: Evm,
+    ) -> eyre::Result<PayloadBuilderHandle<<Node::Types as NodeTypes>::Payload>> {
+        let (tx, mut rx) = mpsc::unbounded_channel();
+
+        ctx.task_executor().spawn_critical("payload builder", async move {
+            #[allow(clippy::collection_is_never_read)]
+            let mut subscriptions = Vec::new();
+
+            while let Some(message) = rx.recv().await {
+                match message {
+                    PayloadServiceCommand::Subscribe(tx) => {
+                        let (events_tx, events_rx) = broadcast::channel(100);
+                        // Retain senders to make sure that channels are not getting closed
+                        subscriptions.push(events_tx);
+                        let _ = tx.send(events_rx);
+                    }
+                    message => warn!(?message, "Noop payload service received a message"),
+                }
+            }
+        });
+
+        Ok(PayloadBuilderHandle::new(tx))
     }
 }
